feat: Rename folder #276
