+++ client_admin/src/pages/App.js
@@ -7,8 +7,8 @@ import LandingPage from "./LandingPage/Index.js";
 import LogInPage from "./LogInPage/Index.js";
 import UserList from "./UserPage/List.js";
 import UserPage from "./UserPage/Index.js";
-import BudgetList from "./BudgetPage/List.js";
-import BudgetPage from "./BudgetPage/Index.js";
+import UBudgetList from "./UserBudgetPage/List.js";
+import UBudgetPage from "./UserBudgetPage/Index.js";
 
 import useAPI from "../hooks/useAPI";
 import useStore from "../hooks/useStore";
@@ -50,8 +50,8 @@ function App() {
           {RouteUser("/DB", <Navigate to="/DB/users" />)}
           {RouteUser("/DB/users", <UserList />)}
           {RouteUser("/DB/users/:_id", <UserPage />)}
-          {RouteUser("/DB/users/:_id/budgets", <BudgetList />)}
-          {RouteUser("/DB/users/:_id/budgets/:budgetId", <BudgetPage />)}
+          {RouteUser("/DB/users/:_id/budgets", <UBudgetList />)}
+          {RouteUser("/DB/users/:_id/budgets/:budgetId", <UBudgetPage />)}
         </Routes>
       </div>
     </Suspense>
